Introduce Distribution.Client.Get.Brancher
