+++ cabal-install/cabal-install.cabal
@@ -232,6 +232,7 @@ executable cabal
         Distribution.Client.Freeze
         Distribution.Client.GenBounds
         Distribution.Client.Get
+        Distribution.Client.Get.Brancher
         Distribution.Client.Glob
         Distribution.Client.GlobalFlags
         Distribution.Client.GZipUtils
